Update window_addList behavior
All but one point that calls window_addlist with a selected item
changed handler calls that handler immediately after.  Update this
to be the default behavior if there's a default item passed in.

The caller in the options window had to be slightly re-worked, but
it turned out to be a cleaner solution.