Removed constraints when adding outfits to a ship.
Change has effected the equip scripts aswell, but that had no advarse
effects.
Made some refractoring in the process, name and comment changes.
The way I build the report is the most lazy way I could think of
without storing the checks in an array/data struct with the appropriate
string, and without writing a custom string writing function.