Replaced FLF intro mission.
The old FLF "pre02" mission is something I'm not fond of. The length
of the mission is ridiculous, you realistically have to land on
Dvaered planets to refuel, and it's ridiculously hard. I also don't
much like the arbitrary restriction on ship choices.

The new mission instead is taking out *one* patrol, ostensibly of
three ships and by yourself, but Dvaered reinforcements arrive.
Shortly after, an event happens:

If you have done missions for the Dvaered before, the leader of the
reinforcements hails you, asking you to defect to them. If you
accept, your score with the Dvaered is made to be no less than 0,
and you are told to go to a certain planet. A few seconds later, a
large group of FLF pilots appears and attacks you. If you refuse,
the same large group of FLF pilots appears and helps you.

If you have not done missions for the Dvaered before, a large group
of FLF pilots appears and helps you.

This new mission actually favors small, fast ships over large ships;
the reinforcements are so powerful that it is usually impossible to
beat them on your own, regardless of the size of your ship. You
have to depend on the help that comes from the FLF fighters.

I am going to recycle the mission you are told this is as a generic
mission for the FLF from the mission computer, accessible only after
this mission is beaten.