Added hidden jump points for the FLF.
There's a problem with the previous setup: the FLF is in a position
that makes what they do impractical at best. As a player, you have
to either have an unreasonably large fuel tank, or land on the
planets of your enemies. This has been remedied by adding hidden
jump points that allow FLF ships to get almost anywhere they would
like within two jumps of a non-enemy planet, and within three jumps
for deeper parts of Dvaered space. Also, it's now possible to go
from the FLF base to the Frontier and back without entering Dvaered
space at all.