Added virtual Sorimid unpresence to Sigur.
It should be noted that virtual unpresence doesn't work; virtual
assets, at least, don't have any effect in the system they're in,
so virtual assets with a range of 0 have no effect at all. But this
is obviously a bug, so once that's fixed, Sigur will have only FLF
ships in it.