Require damage for slowdown when slider is below "enemy presence".
I also added a 2-second delay after time is reset, before speeding
up again, to give you time to react when you take damage. A minor
side effect of this is those times when enemy presence is detected
for one frame, which I think is incorrect (maybe due to the way ships
spawn?), are going to be much more noticeable.