Added a fix for those times when ships are spotted for one frame.
I think that is a result in a defect in spawning of ships. Anyway,
I worked around it by requiring enemy presence for two frames in a
row to trigger time reset. It works as it should. This makes the
system much smoother.

Also made it so the delay for speeding up time again is only used
for damaged-based time resets.