Fixed most outfits being broken by a recent change.
More specifically, I didn't realize that o->stimer could be
"infinity" and that most outfits have it set to this. This was
causing a problem with the formula I was using, understandably
(in real math, the result would be infinity, though apparently the
result in this case was "-nan"; either way, it's a problem). This
has been fixed by only using the new formula if o->stimer is not
INFINITY, and otherwise using the old way of just using the full
cooldown every time.