Added FLF Lancelot and FLF Vendetta.
These ships aren't actually being used by the FLF. Their only
purpose is to be sold to the player at the FLF base. They are
exactly the same, but come with crappy core systems rather than
basic core systems, are slightly cheaper, and don't require a
license to be bought.