Increased the energy usage of the Emergency Shield Booster.
The text describing this outfit claims that energy use is
"disproportionately high", which is complete nonsense. The cost
before was *incredibly* cheap. It has been more than quintupled,
to 160. This amount of energy usage does not render the outfit
useless for small ships, though it does make the energy usage
actually a problem for these ships.