Made Sindbad into a black market planet.
While not as blatant as the pirate worlds, it doesn't make much
sense for Sindbad (the FLF base) to check your licenses when
selling you stuff. While the FLF is certainly more noble than
pirates, it is still an illegal organization, and you're only ever
let into the base at all if you're deemed to be trustworthy.

This doesn't have much if any real effect, because it sells ships
that require no licenses and the only weapon it sells is mace
rockets. This change is more for the principle of it than anything.