Replaced fuel_consumption stat with fuel_consumption characteristic.
fuel_consumption is a flat variable, but ship stats are clearly
meant to be midifiers. This resulted in the game telling the player
that fuel_consumption was a modifier. It's now a (mandatory)
characteristic of the ship instead. This also avoids the problem of
the arbitrary 100 fuel default consumption.

Ships haven't been updated yet. That will be done next commit.

I also removed a bunch of ship info from the shipyard that is
outdated (i.e. characteristics that are no longer used, and that are
now determined by outfits and ship stats).