Reverted the removal of ship info.
I've decided that I'm not in favor of removing this. Naev doesn't
use it, but Naev's engine is very flexible and I think it would be
nice for the engine to be used for other similar games; for this,
having the ability to give specific ships extra stuff can be useful.