Don't make the FLF mad at you just for working for the Dvaered.
It doesn't make sense to do this. Not all patrols involve encounters
with the FLF. Some can have no encounters at all, and some can have
encounters with pirates. It depends on what system you're in. The
FLF should only get mad at you for attacking FLF ships, and the
automatic handling of this should be sufficient for this.