Removed initiating autonav with left-click on planets.
If you're close enough to click on a planet, using Autonav to get
closer to it is basically useless. If anything, it can be confusing
for new players. I changed it to just hail or land; the message
informing you that you're too far away to land is more useful.