Rework and optimize distress
This was meant to be two commits, but the changed lines were adjacent.
The main gameplay change here is that pilots will notice attacks that
happen near them and react appropriately, regardless of whether the
victim has broadcast a distress yet. This should make player piracy in
inhabited systems significantly more difficult.

The second part of this is a few small optimizations to pilot_distress,
only running several of the checks if the player is the attacker.