Added caching for texture transparency maps
When loading a texture that requests transparency mapping, the PNG data
is hashed (MD5). If the cache dir has a file matching the MD5 hash, and
its size is correct, it's loaded instead of generating the transparency
map from scratch.

If a cached map doesn't exist, the generated one will be written to the
cache directory.