Overhaul mouse targeting
The overlay now uses the same click-handling code that everything else
does. Proper fallthrough has been implemented as well, so that a click
on the radar can still target a ship underneath the radar, etc.

Left click is now used for the majority of actions, with the exception
of autonav. Autonav is now exclusively engaged via right-click.