Made it possible to activate outfits that are cooling down.
It makes no sense for a somewhat hot outfit to have to be completely
cool before you can start it up again, and it degrades the usefulness
of outfits with a cooldown period (such as the Emergency Shield Booster)
in a way that's annoying. This way, the cooldown period still limits
the use of such outfits the same amount, but gives the player full
control over when this use is.

I wanted to do this with beam weapons as well, but couldn't figure
out how to do it for them.