Don't lower the player's reputation unless in the faction's territory.
This was something I suggested on IRC to get rid of the problem where
fighting Empire ships that join in fights against you with the Dvaered
negatively affects reputation. I'm not quite sure how well it works.

One particular quirk with this is piracy in uninhabited systems has no
permanent consequences. But it could be that this isn't such a bad
thing.

One problem with this is it doesn't work for the FLF or traders, since
both of these factions are typically encountered nowhere near their
own territory. To solve this, I added a variable indicating whether
the old or new behavior should be used (default new), and set the FLF
and Trader faction to use the old behavior.