Don't offer pirate bounties in the current system.
I always get the same result when I get a pirate bounty in the current
system: I can't find the pirate, and it gets killed by someone else.
I could have solved this by making it spawn close to the planet, but
not doing this was a deliberate measure to prevent the player from
grinding in one system over and over again, so I've just removed the
option for bounties in the current system.