Fix a crash when beams persisted after pilot death
I believe this was due to improper handling in pilot_outfitOff, but
I've added a check for paranoia's sake in pilot_stopBeam as well.

The outfit_* functions used in pilot_outfitOff return -1 for beams,
so a sufficiently-negative stimer turns into a large positive value,
which is presumably what kept the beams going after pilot destruction.