Implement a system for rescuing players that get stranded on load
Since a large number of planets lack a shipyard and outfitter,
renaming or removing core outfits is a dangerous proposition, as
it has the potential to leave players unable to take off without
resorting to save editing.

This script aims to alleviate that. It's not entirely bulletproof, and
it could skip a few cases, but it should reliably get most broken saves
functional again rather painlessly.