Add cross-platform log redirection when not running from a terminal
As --datapath and datapath.lua must be respected, output is initially
copied into memory, and either flushed to disk or deleted once configs
have been loaded.

Redirection is only done is conf.redirect_file is enabled and Naev is
not connected to a terminal.

The files go into ${datapath}/std-{out,err}_YYYY-MM-DD_hh-mm-ss.txt,
and the logs for a session are deleted if stderr is empty.