Initial commit of resizing and fake fullscreen
All of this is exclusive to SDL2.

SDL_WINDOW_FULLSCREEN_DESKTOP allows Naev's window to be the size of
the current screen without setting video mode (and thereby turning off
other screens, messing up background window positioning, etc.), but to
use it, we need to be able to resize the GL viewport on demand.

The work required to allow that also permits dynamic resizing while
playing, as long as a few other things are also reinitialized (a minor
issue).