Add on-the-fly resolution changing via the options menu
SDL2 doesn't allow the resolution to be changed while fullscreen, so
a resolution change while fullscreen triggers the "restart to apply
changes" warning.

It seems that it's fine to go from windowed to fullscreen and back,
even with fullscreen mode setting enabled (does not destroy the GL
context), at least on Linux with Nvidia drivers.

There's also a prompt when changing resolutions or between fullscreen
and windowed, that will restore all settings if no/esc is pressed.