Made the "alive" bounties more doable.
Quite simply, I made the pirates in the "alive" bounty a different
faction, which the pirates consider to be an ally (and so they will
assist them), but which no one considers to be an enemy, and their
only enemy is the player. In effect, no one gets in the way by blowing
them up. Since non-lethal weapons are limited in power, and other
pirates continue to be an obstacle, this doesn't seem to cause the
mission to be overtly too easy.