Added setPlayerStanding Lua function (and a corresponding C function).
Gets rid of having to mod the standing to -standing + whatever you
want. You can now just set the standing straight to what you want.
This is needed for my proposal to change how standing is handled,
but even without that, it's useful in some places (e.g. for cheating,
and for missions like flf_pre02 which either immediately make you an
enemy or immediately make you not an enemy).

Like modPlayerRaw, this function doesn't pass through any Lua code
before setting the standing; it's just a direct modification.