Made outfits automatically start up again when they're done cooling down.
To do this, I got rid of the simplistic "state" variable and replaced
it with a couple attributes indicating whether the outfit is on, and
whether it's cooling down. The result is you can allow the outfit to
automatically come back on much in the same way beams will if you
keep holding the fire button down.