Add some hidden jump points south of Dvaered space
The existing hidden jumps provide no shortcut; they are the same number
of jumps as going through normal jumps. It seems they could serve as a
way to bypass Dvaered space, and these extra jumps help with that. It is
still necessary to pass through one Dvaered system (Amaroq) and the
systems all have a Dvaered presence, but this makes it at least a bit
more realistic for an enemy of the Dvaered, knowing these jumps, to get
between Za'lek and Sirius space.