Changed _fcap_mod_sec into six more specific variables.
Now, there is a different modifier variable for each possible
situation: distress of an enemy, distress of a friend, death of an
enemy, death of a friend, a mission completed by an enemy, and a
mission completed by a friend. For now, I've set all values to 0.3,
the value that _fcap_mod_sec was, with the exception of the variable
for distress of an enemy, which has been set to 0 because there was
previously code which specifically prevented enemy distress from
causing standing gain.

This change makes it possible to do much more fine-tuning to the way
standing gains and losses "leak out" to other factions.