Made the Empire not care about its allies being attacked.
The effect of this is that if the player attacks House Goddard or House
Dvaered, the Empire will not become hostile to the player. Only attacks
on the Empire's own ships will provoke this response.

This change is primarily meant to make it possible to align with the FLF, but
maintain good relations with the Empire.