Cargo Mission Overhaul
Brief Summary:
Cargo delivery missions now only deliver commodities offered at the
origin planet to planets that do not sell that commodity. Jump and
distance rewards depend on the base price of the commodity being
delivered. Rewards are also higher for traversing pirated space.

A new mission was created that asks players to deliver a commodity not
sold on that planet. Pay is dependent on base price of the commodity
and how much is delivered.