Commodity run: check entire system, not just the one planet.
I changed it so that rather than refusing to continue only if the
commodity is available on the current planet, it refuses to continue
if the commodity is available anywhere in the system at all. This
is to avoid situations where you can easily grind in a really boring
way, by just flying back and forth between two planets that are
right next to each other.