Made the cargo hulls give less protection.
There really wasn't much difference in terms of protection between
the cargo hulls and the high-end combat platings. I think this is
a bad design, so I have changed the cargo hulls to all have low-end
protection (equivalent to the standard Unicorp components of the
same size) to make it more of a trade-off.