Removed the map for Sol's location.
The existence of a map suggests that the area is meant to be
explored or has something important in it. Neither is the case for
Sol; the inner nebula is completely impossible to pass through, and
there are no plans to change this in the future (instead, a way
around it is supposed to be added). Considering this, the Sol map
just creates confusion for players and is an expensive, worthless
item.