Added the next FLF Empire alliance campaign mission.
This is where the campaign goes sour. Tense feelings from a minority
of FLF pilots causes some of them to go rogue and attempt to initiate
a coup. The player is then tasked with stopping this coup. Of course,
this is just the beginning.

This mission also unlocks missions similar to the Dvaered patrol
missions, but targeting rogue FLF pilots instead. Unlike the Dvaered
patrol missions, these missions do not give you a reputation boost,
only money, and completion is met with coldness.