Added the next FLF mission: "A Common Enemy"
This is the next mission in the FLF-Empire campaign. Due to some
pirate activity, the leaders of the FLF manage to steer most of the
FLF toward focusing on piracy in the Frontier, and in the process,
distract them from the fact that missions against the Dvaered are
no longer being offered. However, some foreshadowing indicates that
not everyone is convinced.

This mission does the following:

1. Makes the Pirates officially enemies of the FLF.

2. Unlocks more missions where you take out pirates in Frontier space.
   These missions give a reputation boost with the Dvaereds if you're
   near Dvaered systems, so completing them is what is needed to win
   them over.

3. Prevents any anti-Dvaered missions from being offered by the mission
   computer any more.

Next up will be the FLF-Dvaered ceasefire.