Removed the flf_empa campaign.
I've been thinking that this campaign is boring, and makes no real
sense. Plus, I have no intention whatsoever of continuing it. I think
it would be best for the FLF to just stick to terrorism. It's what
they're designed to be, after all.

I've kept around the Lua files for the FLF pirate elimination and
"Rogue FLF" missions. I figure the pirate elimination mission can
be added (in a modified form) as a result of a falling out between
FLF and Pirate factions (after an alliance or deal of some sort),
and there can be "rogue FLF" ships eventually that turn on the FLF
(having been converted by the Empire or Dvaereds) and either try to
escape to spill the location of Sindbad or try to disrupt the FLF's
terrorism on their own. Different plot excuses for those same
missions.