Fixed a mistake in Commodity Run mission.
The Commodity Run mission was printing straight numbers, which gave
ugly and difficult to read results. Fixed; it now uses numstring in
that instance.

I also added "po/*.gmo" to .gitignore, because I noticed that a file
called "po/ja.gmo" was about to be committed. Hopefully that's the
right way to fix it; if not, please let me know.