Added a delay to the creation of commodity runs.
I found that with commodity runs constantly appearing on certain
planets, I felt compelled to keep doing commodity runs for the same
planet over and over again. That's fine in and of itself, but the
problem is that this is not a very fun way to play; players should
be encouraged to explore, not just visit the same planet over and
over again. To address this, I've added a 7 day delay
between the acceptance, completion, or cancellation of a commodity
run and when a new commodity run is allowed to be created. This is
such an amount that it doesn't limit the number of commodity runs
you can do realistically, but it keeps you waiting just long enough
to discourage you from doing a run for the same planet over and over.

Also added "po/Makefile.in.in~" to .gitignore, since when I committed
this it added that file as well.