Started on flf_dvk02.
This is the "pirate alliance" mission. The FLF is starting to get
its suppliers cut off and needs a new source. Luckily, it is believed
(correctly) that a pirate stronghold is nearby, so Benito sends you
to the approximate location the FLF thinks the pirates are located
to investigate and attempt to secure trade relations. The end result
will be to make the FLF and Pirate factions into allies... that is,
until a fallout between them occurs some time after, forcing the FLF
to deal with piracy in their own territory.

But for this mission, we're just establishing the alliance. What could
possibly go wrong?