Added support for loading arbitrary images as ship graphics.
Three new possible tags have been added to ship XML files:
"gfx_space", "gfx_engine", and "gfx_comm". These allow you to define
ship graphics outside of the normal ship graphics tree. I've tested
the new tags, and there are no problems; ships with graphics defined
in this way load just the same as with the GFX tag (which is still
preferred for most ships, for the sake of brevity).