Fixed a bug that allowed you to buy maps with no new information.
The bug was caused when there were destroyed assets within the map.
Whether or not the asset actually existed was not being checked,
so you could always buy the map again and again, wasting your
credits. Fixed by considering any planet that either doesn't exist
or is outside of the planetname_stack to be "known". This fixes the
problem and doesn't seem to have any unintended side-effects.

Big thanks to both bobbens and BTAxis for helping me navigate the
code on this one.