Added support for factions not using generic NPC messages.
This is so that the Thurion, Proteron, FLF, and Pirate factions
don't say things that make no sense for them. It also avoids the
strange presence of non-FLF civilians in Sindbad and non-pirate
civilians in Pirate strongholds.

Mission aftercare messages and lore text are affected. Hint texts
are left as-is (there's no potential continuity problem with those).