Changed "STU", "STP", and "SCU" to "seconds", "periods", and "cycles".
This is meant for better readability; it makes the different units
more distinguished from each other.

"STU" was replaced with "seconds" instead of "units" for two reasons:
one, "units" is too generic (it's used for all kinds of measurements,
not just time measurements), making it a confusing name. Two, the SCU
is the same as a second in-universe anyway, and I don't see the
general population abandoning that term.

For the auto generation of "pretty" time text, I replaced e.g "12 STU"
with "12s", and e.g. "8.32 STP" with "8.32p".

Also fixed some language mistakes in dialogs I happened to notice,
plus some very weird usage of the time units that probably resulted
from a misunderstanding of how long they're supposed to be (there
was a father concerned because his daughter disappeared several
*seconds* ago, a food station that had a harvest every day, and a
weekly mining trip to one planet that was talked about as if that
wasn't frequent).

Just a thought that just came to mind as I was writing this: perhaps
it would be better to call decaperiods "daps"? Not truly necessary,
but it would be an interesting idea because it's a name for a unit
that kind of look like "days", while also serving as a shorthand
for the full official name. Any thoughts?