Reverted the range increase of the Electron Energy Cell from before.
Commit 04dd81706c75a451894dc16eacd81845f02f1372 increased the range
of this weapon, used by the Collective Heavy Drone, dramatically.
The range was so much that these drones could shoot you from halfway
across the system, meaning that their presence was pretty much a
guaranteed death sentence (especially given how powerful this weapon
is). This is simply a reversion of that range increase while retaining
all the other buffs the weapon got.

The result is still extremely powerful, but survivable. It also shouldn't
have any real effect on usability by players since using an unguided
bullet at such a large range is impractical unless you have a whole
swarm on your side (which, as a player, you never do).