Increased the size of the Brushed OSD, made its height dynamic.
I originally wanted to make its width dynamic as well, but the
way the OSD is set makes that not really work so well. I'm not really
confident in my ability to fix the OSD code to make it more dynamic
in this fashion, so for now I've just set it to a wider static OSD
width.