Several improvements to the Sindbad base offensive mission.
The way this mission was structured was quite dated. I made several
changes to fix problems and make it work better overall.

- Instead of timing, the phases are determined by how many FLF you've killed.
- Uses Lancelot and Vendetta for FLF instaed of Vendetta and Ancestor.
- No longer enters cinematic mode in the final phase.
- No longer controls FLF ships; they just do what the AI wants.

The only thing that need improvement now is that the FLF just don't
chase the group effectively. Need to come up with a solution to that.
(It's not as bad as it was previously though.)