Added variable "time mod" bases.
This was a pretty simple idea I had: bigger ships see time passing
faster. So in effect, by using a big ship, there's a trade-off: it's
stronger, sure, but it's also harder to react.

Justification: the bigger ships are harder to pilot and therefore
it takes longer to send out the commands.

Also made both the Collective Drone and the Hyena have *slower*
time passage. This gives them an advantage over heavier craft: with
slower time passage, the player has more time to react.