Substantially buffed the Hyena's agility.
The turn mod has been increased from 10% to 15%, and the thrust mod
has been increased from 10% to 100% (i.e. double thrust). This makes
dodging bullets, the intended advantage of the Hyena, much easier,
which in turn makes the ship more competitive. Also has a nice
side-effect of making pirates a bit more of a threat.