New "bounty" mission for pirates
I've replaced the existing "empire pirate bounty" mission with a
variation of the "Dead Or Alive" bounty mission. It is, however,
"kill only" and renamed to "Assassination Job". It shows up in the
mission computer and is available through the mission computer on
Independent planets, pirate strongholds, and the FLF base.  Requires
a mercenary license unless on a black market world (i.e. Pirate
strongholds or the FLF base).

I might want to expand this to include other major factions in the
future. For now, though, it is limited to eliminating Empire ships.

Hitman 3 has been disabled because it is now obsolete, and the
missions that previously required a Pirate landing pass now only
require positive reputation with the pirates.

This commit also includes one more small change: Hyenas are now
counted as pirate ships for the purpose of the pirate reputation
loss event. The reasons for this are (1) because the pirates are
often seen with these ships, and (2) to give the player access to
something considered to be a pirate ship early on.