Removed ship location altogether.
This was discussed on #719: ship swapping really doesn't add anything
to the game, so this simply removes that feature. Ships no longer have
a location, they just travel around the galaxy with you like outfits
do.

I believe this probably also fixes #719, though it's hard to tell
for sure. The reason I think it might have fixed it is because that
bug seemed to be most easily triggered by following a transport
immediately with a swap. I suspect that somewhere the copying of the
"loc" string of ships was going wrongly, so if that is indeed the
case, the fact that that variable has been entirely removed should
fix it. This is just speculation, though, and it's impossible to
tell for sure.

The NPC tip about having to transport ships has also been removed,
since it is no longer true.