Added a pirate version of the patrol mission.
It doesn't actually work right now, not because there's anything wrong
with the mission, but because there just happens to be no system near
any pirate stronghold with both sufficient enemy presence and a
sufficient number of planets. But should that change in the future,
this mission will become available.