Upped the rewards on most unique missions, plus other minor improvements.
Whew, that's a lot!

So the gist of this is that most unique missions have had their
rewards substantially increased. The general rule of thumb I followed
was to make the rewards 10 times what they previously were, but of
course there were exceptions to this. Most notably, most of the
Collective campaign previously had no rewards at all, so very large
rewards (progressing slowly from 500k to 5m) have been added to it.
Some, on the other hand, were only increased slightly.

The main goal of this is to balance out the mission rewards, which
has been a serious problem with Naev. It used to be that standard
missions gave you more money than unique missions, and this change
should hopefully reverse that and encourage players to try more
missions.

While I was doing that, I also fixed the presention of rewards and
numbers for several missions, and corrected some small dialog problems
I found (mostly typos).