Removed player_getRating and all uses of it.
Inspired by something @bobbens said in #832 that I agree with
("I don't think anybody would miss it if removed").  I didn't go
all-out and remove the combat rating mechanic because I don't think
its existence alone is a problem as long as it's just cosmetic (and
it might even be a nice little nod to EV and a way to sort of get
a feel for how much combat you've done this game), just actual use
by Lua code, which was rare.

Missions that previously modified their difficulty based on this no
longer do so. In particular:

* emp_hcvl now always gives you an "easy" ship.
* trader_escort now always gives you the full range of options for convoy size.

Also worth noting that dat/missions/common.lua has been removed both
because no Lua code was actually using it, and because everything
there seems to be centered around the combat rating mechanic.
hitman3's Lua file has also been removed because it is no longer in
use (due to being obsolete) and doesn't really have anything that
needs to be saved. (Both of these decisions were a response to the
use of player_getRating() in these files.)

Of course, all events and missions that had <cond> attributes with
getRating() have had those getRating() calls removed.