Added "Coming Out" mission.
This is the first in a series of missions about the story of a
transgender person coming out and becoming a pilot. It's in the
Soromid faction because a later mission is going to involve getting
medical treatment from the Soromid. This mission just serves to
introduce the character (Chelsea) and set up later missions with a
few hooks.

I also fixed a couple of small issues that I happened to notice as
I was working on the new mission: a typo in flf_dvk07 ("hay" instead
of "hey"), and an autogenerated file (po/stamp-po) getting included
by git add --all.