Auto add / remove missiles, plus set missile costs to 0
Just the start of the work here. This causes it to be unnecessary
to buy missiles. They remain in the shop at cost 0 for the time
being (and do nothing). Next up is removing those missiles from the
shops, and consolidating the info about missiles and their launchers
into the launchers themselves. Missiles will stay separate outfits,
so I'm thinking of just adding a button to the launcher to view the
missile's info, or otherwise I might just programmatically paste the
missile info into the launcher somehow. Undecided.