Modified missiles to make guided missiles medium.
This is an attempt to redesign the way missiles are balanced: guided
missiles of all kinds have been redefined as medium (weights adjusted
accordingly), and the Banshee rocket has been redefined as small. The
idea here is to restrict guided missiles to bombers and up, mainly for
use by bombers, corvettes, and light destroyers.

This of course will cause equipping of AI fighters to stop working
correctly. I will address that in the next commit.