Fix mission "The Runaway"
During the mission player gets to decide how to finish the mission:
they can release or capture Cynthia.
Player's choice is not saved and doesn't affect the outcome of the
mission if the game is reloaded before the mission is finished.

The fix is to check the text of the OSD item in the final part of the
mission:

* Implement misn.osdGetActiveItem
* Check active OSD item to determine mission state