Perhaps fixed a rare segfault, maybe
I can't confirm whether or not what I did fixed the problem, but I do
know that it segfaulted at the "o = pilot->outfits[i]->outfits;" line,
and I know that "pilot" is not NULL. So my best guess is that maybe
pilot->outfits[i] is NULL sometimes? So I added a check for that.
The segfault was extremely rare so I can't be 100% sure if I've solved
it or not.