Increased trader escort mission reward.
The trader escort mission is probably the most tedious regular mission
in the game, so I think it's best for it to be a very rewarding one.
Not so rewarding that it's a must-do, but rewarding enough that there's
a point to doing it instead of a bounty hunt.

I've also decreased the difference between small and large convoys.
It's still there, but not as drastic. Previously the largest was 20x
the distance reward of the smallest; now it's only 10x.