Add a small world of Proteron survivors in Iris
This consists of the survivors of the Proteron who were left on this
side of the nebula.

According to the lore, the Proteron were a Great House before the
incident. Since they were not a secret, unlike the Thurion, there
probably should be some mention of them. It seems logical that some
would have been stranded on this side of the nebula.

This provides some interest in the Proteron for the player, and could be
a useful site for future missions (a Proteron invasion might begin by
secretly establishing contact with and control of Osiris).

https://web.archive.org/web/20170813171731/http://wiki.naev.org/wiki/Proteron