Replace opengl_ext.c with files generated using Galogen
Looking at several OpenGL extension loading libraries, I like Galogen
because it allows generating files for any choice of GL version and
extensions. So when we (hopefully) switch to GL 3.2 Core, we can
generate headers for such and it will be a compile error to use features
that are removed in that version, or added by later versions.

Generated with:
```
galogen gl.xml --ver 3.0 \
               --profile compatibility \
               --exts ARB_texture_filter_anisotropic
```

https://github.com/google/galogen